+++ src/app/app.component.ts
@@ -1,10 +1,11 @@
 import { Component } from '@angular/core';
+import {HeaderComponent} from "./header/header.component";
 
-@Component({
-  selector: 'app-root',
+@Component({ // dekorator konfigurujący naszą klasę
+  selector: 'app-root', // na jaki markup ma nasza klasa być wywoływana
   standalone: true,
-  imports: [],
-  templateUrl: './app.component.html',
-  styleUrl: './app.component.css',
+  imports: [HeaderComponent],
+  templateUrl: './app.component.html', // html naszego komponentu
+  styleUrl: './app.component.css', // style naszego komponentu
 })
 export class AppComponent {}