+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 ## 02.01.2025
-- Project initialization
+- New header component
 - Project initialization